Add missing java docs, and change imports related to FHIR module due to previous refactoring
