+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/tests/facility/ViewFacilityComplexTests.java
@@ -3,8 +3,8 @@ package ca.bc.gov.health.qa.autotest.plr.web.tests.facility;
 import ca.bc.gov.health.qa.autotest.plr.data.ViewFacilityConstants;
 import ca.bc.gov.health.qa.autotest.plr.data.ViewProviderConstants;
 import ca.bc.gov.health.qa.autotest.plr.fhir.FHIRController;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.FacilityMaintainConfig;
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainFacilityBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.facility.FacilityMaintainConfig;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility.MaintainFacilityBuilder;
 import ca.bc.gov.health.qa.autotest.plr.util.UserType;
 import ca.bc.gov.health.qa.autotest.plr.web.actions.facility.ViewFacilityActions;
 import ca.bc.gov.health.qa.autotest.plr.data.SearchFacilityConstants.*;
